Added in SINR, PropFair to Clients
SINR and PropFair functions have been added to clients. Currently, the mean SINR is set to 15, with std =2.5. Round std are set to 1. This will keep virtually all values been 0 and 25.

A DataQuality function has also been added, but it only produces a flat score of 20. This is to simply get the CIF adding SINR and DataQuality together.